integrate new data classes and GUI

+++ components/projects.py
@@ -2,7 +2,7 @@
 from __future__ import annotations
 import os, glob, json,sys
 from copy import deepcopy
-from components.data import ImageInfo, LabelInfo, Labels
+from components.data import BBox, ImageInfo, LabelInfo, Labels
 sys.path.append("./")  # add ROOT to PATH
 import custom_utils
 
@@ -28,20 +28,40 @@ class Project(object):
             ext_ann_path = p["ext_annotations"]
             self.import_annotations(data_path, ext_ann_path)
         
+        
     def import_annotations(self, data_path, yolo_annotations_path):
 
         for f in glob.glob(f"{yolo_annotations_path}/*.txt"):
             name = os.path.basename(f).rsplit('.')[0]
+
             img_path = f"{data_path}/{name}.jpg"
-            img_size = custom_utils.get_image_size(img_path)
-            bboxes = custom_utils.import_yolo_predictions(f, img_size[0], img_size[1])
             img_info = ImageInfo(name, img_path)
+            bboxes = custom_utils.import_yolo_predictions(f, img_info.w, img_info.h)
 
             img_info.add_bboxes(bboxes)
-            self.imgs[img_path] = img_info
+            self.imgs[name] = img_info
+            
+            if not os.path.exists(f"projects/{self.name}/annotations/{name}.json"):
+                bboxes_list = list(map(lambda x: x.as_obj(), bboxes))
+                with open(f"projects/{self.name}/annotations/{name}.json", "w") as fp:
+                    data = {"img_path": img_path, "bboxes": bboxes_list}
+                    json.dump(data, fp, indent=1)
+    
+    def load_annotations(self, ):
 
-            with open(f"projects/{self.name}/annotations/{name}.json", "w") as fp:
-                json.dump(bboxes, fp, indent=1)
+        for f in glob.glob(f"projects/{self.name}/annotations/*.json"):
+            name = os.path.basename(f).rsplit('.')[0]
+            with open(f, "r") as fp:
+                data = json.load(fp)
+            
+            img_path = data["img_path"]
+
+            bboxes = list(map(lambda x: BBox(x["xmin"],x["ymin"],x["xmax"],x["ymax"], x["label"], x["conf"]), data["bboxes"]))
+            img_info = ImageInfo(name, img_path)
+            
+            img_info.add_bboxes(bboxes)
+            self.imgs[name] = img_info
+        
 
     def get_image(self):
         for img in self.imgs:
@@ -50,8 +70,8 @@ class Project(object):
     def load_labels(self):
 
         labels = Labels()
-        for k in self.labels:
-            obj = self.labels[k]
+        for k in self.info_obj["labels"]:
+            obj = self.info_obj["labels"][k]
             l = LabelInfo(
                 obj["index"], 
                 obj["label"],
@@ -93,8 +113,9 @@ class Project(object):
                 with open(f"projects/{self.name}/annotations/{img_info.name}.json", "w") as fp:
                     scaled_bboxes = []
                     for bbox in img_info.bboxes:
-                        scaled_bboxes.append(bbox.scale((img_info.scaled_w, img_info.scaled_h), (img_info.w, img_info.h)))
-                    json.dump(scaled_bboxes, fp, indent=1 )
+                        scaled_bboxes.append(bbox.scale((img_info.scaled_w, img_info.scaled_h), (img_info.w, img_info.h)).as_obj())
+                    data = {"img_path": img_info.path, "bboxes": scaled_bboxes}
+                    json.dump(data, fp, indent=1 )
                 img_info.set_changed(False)
 
 
@@ -102,14 +123,12 @@ def load_projects():
 
     projects = []
     for p in glob.glob("projects/*"):
-        # print(os.path.basename(p))
+        
         with open(p + "/info.json", "r") as fp:
             info_obj = json.load(fp)
         projects.append(Project(os.path.basename(p), info_obj))
     
     return projects
-
-
 
 
 if __name__ == "__main__":
